Added testing plugins to pluginManagement
Added surefire-plugin specific version (the default is 2.12.4, which doesn't pick up the jupiter tests), and specified jacoco-maven-plugin with a common configuration to pluginManagement.
It's up to the projects using this parent pom to actually use these plugins when applicable.
Also added common configuration for the spring-boot-mavn-plugin, to save some more repetition.