Added plugin configuration to help with manifest info (#3)
With these plugins, the MANIFEST.MF in the resulting jar should end up with version information.
Using maven-jar-plugin, the version should be added to the manifest.
Using buildnumber-maven-plugin, a datetime and revision part can be added to the version as well. This does require some extra plugin configuration in the project using this parent pom: can't enable that by default as it relies on SCM information.