Avoid issue in IDE with javaagent not working (#49)
When running unittests in IDE (Eclipse at least), as it was it would run into an error and I'd manually have to change the run configuration.
This approach does not rely on maven-dependency-plugin to get the correct jar, and so works straight away in IDE with the correct java agent.